Fixed one more case
